Move stake state / instructions into solana_program
