+++ programs/stake/src/lib.rs
@@ -11,6 +11,3 @@ solana_sdk::declare_id!("Stake11111111111111111111111111111111111111");
 pub fn add_genesis_accounts(genesis_config: &mut GenesisConfig) -> u64 {
     config::add_genesis_account(genesis_config)
 }
-
-#[macro_use]
-extern crate solana_frozen_abi_macro;
