+++ sdk/program/src/stake/mod.rs
@@ -1,0 +1,5 @@
+pub mod config;
+pub mod instruction;
+pub mod state;
+
+crate::declare_id!("Stake11111111111111111111111111111111111111");
